refactor by moving uploadToS3 function to common
