+++ packages/common/src/utils.ts
@@ -1,0 +1,13 @@
+export const uploadToS3 = async (url: string, blob: Blob) => {
+	try {
+		const response = await fetch(url, {
+			method: 'PUT',
+			body: blob,
+		});
+		const status = response.status;
+		return status === 200;
+	} catch (error) {
+		console.error('upload error:', error);
+		return false;
+	}
+};
