+++ packages/worker/src/index.ts
@@ -5,24 +5,19 @@ import {
 	parseTranscriptJobMessage,
 	isFailure,
 	deleteMessage,
-	getFile,
-	getS3Client,
 	changeMessageVisibility,
-	TranscriptionConfig,
 } from '@guardian/transcription-service-backend-common';
-import type {
-	OutputBucketUrls,
-	TranscriptionOutput,
-} from '@guardian/transcription-service-common';
+import { type TranscriptionOutput } from '@guardian/transcription-service-common';
 import { getSNSClient, publishTranscriptionOutput } from './sns';
 import {
 	getTranscriptionText,
 	convertToWav,
 	getOrCreateContainer,
-	Transcripts,
 } from './transcribe';
 import path from 'path';
+
 import { updateScaleInProtection } from './asg';
+import { getFileFromS3, uploadAllTranscriptsToS3 } from './util';
 
 const main = async () => {
 	const config = await getConfig();
@@ -136,59 +131,6 @@ const main = async () => {
 		}
 	} finally {
 		await updateScaleInProtection(region, stage, false);
-	}
-};
-
-const getFileFromS3 = async (config: TranscriptionConfig, s3Key: string) => {
-	const s3Client = getS3Client(config.aws.region);
-
-	const file = await getFile(
-		s3Client,
-		config.app.sourceMediaBucket,
-		s3Key,
-		config.app.stage === 'DEV' ? `${__dirname}/sample` : '/tmp',
-	);
-
-	return file;
-};
-
-export const uploadAllTranscriptsToS3 = async (
-	destinationBucketUrls: OutputBucketUrls,
-	files: Transcripts,
-) => {
-	const getBlob = (file: string) => new Blob([file as BlobPart]);
-	const getFileName = (file: string) => path.basename(file);
-	const blobs: [string, string, Blob][] = [
-		[getFileName(files.srt), destinationBucketUrls.srt, getBlob(files.srt)],
-		[getFileName(files.json), destinationBucketUrls.json, getBlob(files.json)],
-		[getFileName(files.text), destinationBucketUrls.text, getBlob(files.text)],
-	];
-
-	try {
-		for (const blobDetail of blobs) {
-			const [fileName, url, blob] = blobDetail;
-			const response = await uploadToS3(url, blob);
-			if (!response) {
-				throw new Error(`Could not upload ${fileName} to S3`);
-			}
-			console.log(`Successfully uploaded ${fileName} to S3`);
-		}
-	} catch (error) {
-		console.error('failed to upload transcript to S3', error);
-	}
-};
-
-const uploadToS3 = async (url: string, blob: Blob) => {
-	try {
-		const response = await fetch(url, {
-			method: 'PUT',
-			body: blob,
-		});
-		const status = response.status;
-		return status === 200;
-	} catch (error) {
-		console.error('upload error:', error);
-		return false;
 	}
 };
